server/requirements-lock.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==5.0.0
bleach==6.2.0
boolean.py==5.0
CacheControl==0.14.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.3
filelock==3.19.1
frozenlist==1.7.0
gotrue==2.9.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
license-expression==30.4.4
limits==5.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.4
packageurl-python==0.17.5
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.4.0
pluggy==1.6.0
postgrest==0.17.0
propcache==0.3.2
psutil==7.1.0
psycopg2-binary==2.9.10
py-serializable==2.1.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
qrcode==8.2
realtime==2.20.0
requests==2.32.5
rich==14.1.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
starlette==0.48.0
storage3==0.8.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.20.0
supabase-functions==2.20.0
supafunc==0.6.0
toml==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
